fix(fs-mutation): platform-correct Mode cast width + useless conversion
nix::sys::stat::Mode wraps libc::mode_t, which is u32 on Linux but u16
on macOS/BSD. set_permissions.rs's chmod helper previously assumed one
width; switched to an inferred `as _` cast with a platform-conditional
clippy expect so both widths compile cleanly and cast correctly.

preflight.rs's disk-space guard had a redundant u64::from() wrapping a
value nix::sys::statvfs already returns as u64 on this platform --
first surfaced by clippy running on Linux for the first time.

+++ crates/substrate-fs-mutation/src/preflight.rs
@@ -48,7 +48,7 @@ fn check_disk_space_sync(path: &Path, required_bytes: u64) -> SubstrateResult<()
         correlation_id: None,
     })?;
 
-    let available: u64 = u64::from(stat.blocks_available()) * stat.block_size();
+    let available: u64 = stat.blocks_available() * stat.block_size();
     let needed = required_bytes.saturating_add(FREE_SPACE_CUSHION_BYTES);
 
     if available < needed {
